+++ VecMathREPLCpp/vecmathlib/VecMathListener.h
@@ -72,8 +72,8 @@ private:
 	std::stack<IMatrix*> m_ExprStack;
 	std::string m_CurrentCodeLine;
 	std::string m_HelpString;
-	static constexpr float m_DegToRad{ M_PI / 180 };
-	static constexpr float m_RadToDeg{ 180 / M_PI };
+	static constexpr float m_DegToRad{static_cast<float>( M_PI / 180 ) };
+	static constexpr float m_RadToDeg{ static_cast<float>(180 / M_PI )};
 
 	JokeGenerator m_JokeGenerator;
